Enable coverage at runtime
This is the most infuriating bug yet.
Because pytest-asyncio-cooperative was installed in the same environment as pytest it was blocking all coverage. There was no need to enable it, the simple act of installing it was enough to completely block any attempts at trying to check coverage of any test. (Even synchronous ones!)
I will never recover from this absolue bs.